Run tests using test data in test_data.py
This avoids having to load card data from disk or network in order to
run tests. The test data is a selection of 10 randomly chosen cards from
the netrunnerdb dataset converted into the format we expect for card
database entries.

This also decouples tests from netranker.utils, which is now only used
to manage card data on the cli.